+++ govcd/system_test.go
@@ -334,3 +334,75 @@ func (vcd *TestVCD) Test_QueryOrgVdcNetworkByName(check *C) {
 	check.Assert(orgVdcNetwork[0].Name, Equals, vcd.config.VCD.Network.Net1)
 	check.Assert(orgVdcNetwork[0].ConnectedTo, Equals, vcd.config.VCD.EdgeGateway)
 }
+
+func (vcd *TestVCD) Test_QueryProviderVdcEntities(check *C) {
+	providerVdcName := vcd.config.VCD.ProviderVdc.Name
+	networkPoolName := vcd.config.VCD.ProviderVdc.NetworkPool
+	storageProfileName := vcd.config.VCD.ProviderVdc.StorageProfile
+	if providerVdcName == "" {
+		check.Skip("Skipping Provider VDC query: no provider VDC was given")
+	}
+	providerVdcs, err := vcd.client.QueryProviderVdcs()
+	check.Assert(err, IsNil)
+	check.Assert(len(providerVdcName) > 0, Equals, true)
+
+	providerFound := false
+	for _, providerVdc := range providerVdcs {
+		if providerVdcName == providerVdc.Name {
+			providerFound = true
+		}
+
+		if testVerbose {
+			fmt.Printf("PVDC %s\n", providerVdc.Name)
+			fmt.Printf("\t href    %s\n", providerVdc.HREF)
+			fmt.Printf("\t status  %s\n", providerVdc.Status)
+			fmt.Printf("\t enabled %v\n", providerVdc.IsEnabled)
+			fmt.Println("")
+		}
+	}
+	check.Assert(providerFound, Equals, true)
+
+	if networkPoolName == "" {
+		check.Skip("Skipping Network pool query: no network pool was given")
+	}
+	netPools, err := vcd.client.QueryNetworkPools()
+	check.Assert(err, IsNil)
+	check.Assert(len(netPools) > 0, Equals, true)
+	networkPoolFound := false
+	for _, networkPool := range netPools {
+		if networkPoolName == networkPool.Name {
+			networkPoolFound = true
+		}
+		if testVerbose {
+			fmt.Printf("NP %s\n", networkPool.Name)
+			fmt.Printf("\t href %s\n", networkPool.HREF)
+			fmt.Printf("\t type %v\n", networkPool.NetworkPoolType)
+			fmt.Println("")
+		}
+	}
+	check.Assert(networkPoolFound, Equals, true)
+
+	if storageProfileName == "" {
+		check.Skip("Skipping storage profile query: no storage profile was given")
+	}
+	storageProfiles, err := vcd.client.QueryProviderVdcStorageProfiles()
+	check.Assert(err, IsNil)
+	check.Assert(len(storageProfiles) > 0, Equals, true)
+	storageProfileFound := false
+	for _, sp := range storageProfiles {
+		if storageProfileName == sp.Name {
+			storageProfileFound = true
+		}
+		if testVerbose {
+			fmt.Printf("SP %s\n", sp.Name)
+			fmt.Printf("\t enabled     %12v\n", sp.IsEnabled)
+			fmt.Printf("\t storage     %12d\n", sp.StorageTotalMB)
+			fmt.Printf("\t provisioned %12d\n", sp.StorageProvisionedMB)
+			fmt.Printf("\t requested   %12d\n", sp.StorageRequestedMB)
+			fmt.Printf("\t used        %12d\n", sp.StorageUsedMB)
+			fmt.Println("")
+		}
+	}
+	check.Assert(storageProfileFound, Equals, true)
+
+}