+++ govcd/system_unit_test.go
@@ -8,7 +8,6 @@
 package govcd
 
 import (
-	"os"
 	"testing"
 )
 
@@ -105,7 +104,7 @@ func Test_BareEntityID(t *testing.T) {
 			t.Fail()
 		}
 		if bareId == it.expected {
-			if os.Getenv("TEST_VERBOSE") != "" {
+			if testVerbose {
 				t.Logf("ID '%s': found '%s' as expected", it.rawId, it.expected)
 			}
 		} else {
@@ -120,7 +119,7 @@ func Test_BareEntityID(t *testing.T) {
 			t.Fail()
 		}
 		if bareId == it.expected {
-			if os.Getenv("TEST_VERBOSE") != "" {
+			if testVerbose {
 				t.Logf("ID '%s': found '%s' as expected", it.rawId, it.expected)
 			}
 		} else {
@@ -215,7 +214,7 @@ func Test_GetUuidFromHref(t *testing.T) {
 			t.Fail()
 		}
 		if bareId == it.expected {
-			if os.Getenv("TEST_VERBOSE") != "" {
+			if testVerbose {
 				t.Logf("ID '%s': found '%s' as expected", it.rawHref, it.expected)
 			}
 		} else {
@@ -230,7 +229,7 @@ func Test_GetUuidFromHref(t *testing.T) {
 			t.Fail()
 		}
 		if bareId == it.expected {
-			if os.Getenv("TEST_VERBOSE") != "" {
+			if testVerbose {
 				t.Logf("ID '%s': found '%s' as expected", it.rawHref, it.expected)
 			}
 		} else {